move: Change dir structure
- images 파일은 src 보단 public에서 관리하는게 맞다고 생각
- /UI 가 /components 안에 있어야 함 (실수)
