+++ public/index.html
@@ -2,11 +2,11 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" type="image/png" href="../src/images/Logo.png" />
+    <link rel="icon" type="image/png" href="./images/Logo.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta name="description" content="MOIDA" />
-    <link rel="apple-touch-icon" href="../src/images/Logo.png" />
+    <link rel="apple-touch-icon" href="./images/Logo.png" />
     <link rel="manifest" href="./manifest.json" />
     <title>MOIDA</title>
   </head>
